requirements.txt
# API and HTTP
requests==2.31.0
urllib3==2.1.0

# Data handling
python-dotenv==1.0.0

# CLI and terminal UI
rich==13.7.0
tqdm==4.66.1

# Date and time handling
python-dateutil==2.8.2

# File operations
pathlib==1.0.1

# Rate limiting
ratelimit==2.2.1

# Logging
colorlog==6.8.0